requirements.txt
pyftpdlib
pysendfile
circus
gearman
Django
sh
MySQL-python
djangorestframework
markdown
pyyaml
django-filter
